Z8G4 full checkpoint: 133x speedup breakthrough, complete state dump
The Z8 was running at 0.75% capacity. Software fixes alone gave 133x:
- Governor: powersave → performance
- Threads: 64 → 32 (HT kills AVX-512)
- IPEX optimize
- KMP/OMP settings

32B forward: 210s → 1.57s
50K fine-tuning steps: 365 days → 3 days

Checkpoint includes: all optimization findings, manifold fingerprints
(15 models), thermostat results, holographic architecture results,
pipeline status, hardware profile, and next steps.

+++ z8_pipeline_32b/pipeline.py
@@ -222,7 +222,7 @@ def main():
     print(f"\n{'='*60}\n=== Loading {args.model} ===\n{'='*60}")
     t0 = time.time()
     model = AutoModelForCausalLM.from_pretrained(
-        args.model, dtype=dtype, low_cpu_mem_usage=True,
+        args.model, torch_dtype=dtype, low_cpu_mem_usage=True,
         trust_remote_code=True, attn_implementation="eager").to(device).eval()
     L = model.config.num_hidden_layers
     d = model.config.hidden_size
